use nextDouble
git-svn-id: svn+ssh://svn.cprover.org/srv/svn/cbmc/trunk@2572 6afb6bc1-c8e4-404c-8f48-9ae832c5b171

+++ regression/cbmc-java/double1/double1.java
@@ -4,7 +4,7 @@ class double1
   {
     java.util.Random random = new java.util.Random(42);
     
-    double d=random.nextFloat();
+    double d=random.nextDouble();
 
     // conversions
     int i=(int)d;
